chore(zero-cache): structure rm-vs-load benchmark
