+++ packages/zero-cache/bench/rm-vs-load/e2e.ts
@@ -36,4 +36,4 @@ console.log(
   ].join('\n'),
 );
 
-await import('./rm-vs-load.ts');
+await import('./index.ts');
