+++ packages/zero-cache/bench/rm-vs-load/perf-utils.ts
